feat: compose ui snapshot test 방식 변경
- test에서 paparazzi -> androidTest에서 local screenshot test
