--- core/src/androidTest/kotlin/team/duckie/quackquack/core/component/README.md
@@ -1,3 +1,0 @@
-## androidTest/component
-
-이 모듈은 UI 컴포넌트의 로직만을 테스트합니다. UI 컴포넌트의 비쥬얼 테스트는 [스냅샷 테스팅](https://github.com/cashapp/paparazzi)으로 `test/component`에서 진행됩니다.
