+++ core/src/androidTest/kotlin/team/duckie/quackquack/core/component/TextTest.kt
@@ -7,13 +7,17 @@
 
 package team.duckie.quackquack.core.component
 
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.text.BasicText
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTouchInput
 import androidx.compose.ui.text.SpanStyle
+import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.text.withStyle
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Rule
 import org.junit.Test
@@ -28,14 +32,20 @@ import strikt.assertions.withNotNull
 import team.duckie.quackquack.core.material.QuackColor
 import team.duckie.quackquack.core.material.QuackTypography
 import team.duckie.quackquack.core.util.Empty
+import team.duckie.quackquack.core.util.captureToBitmap
 import team.duckie.quackquack.core.util.immediateXMoveTo
-import team.duckie.quackquack.core.util.onQuackNodeWithTag
+import team.duckie.quackquack.core.util.markGolden
+import team.duckie.quackquack.core.util.markTest
+import team.duckie.quackquack.core.util.onGolden
+import team.duckie.quackquack.core.util.onTest
+import team.duckie.quackquack.core.util.screenshotTest
 import team.duckie.quackquack.core.util.setQuackContent
 
 /**
  * [QuackTextScope]의 로직을 테스트합니다.
  *
  * - `Modifier.span`과 `Modifier.highlight`는 같이 사용할 수 없음
+ * - `Modifier#span`과 `buildAnnotatedString`은 같은 텍스트 비쥬얼을 그림
  * - `Modifier.highlight`로 특정 텍스트에 [SpanStyle] 및 onClick 이벤트를 설정할 수 있음
  * - `Modifier.highlight`로 특정 텍스트에 [SpanStyle] 및 광역 onClick 이벤트를 설정할 수 있음
  */
@@ -46,7 +56,7 @@ class TextTest {
 
     // - `Modifier.span`과 `Modifier.highlight`는 같이 사용할 수 없음
     @Test
-    fun Cannot_Use_ModifierSpan_And_ModifierHeight_Same_Time() {
+    fun cannot_use_ModifierSpan_and_ModifierHeight_same_time() {
         expectThrows<IllegalStateException> {
             composeTestRule.setQuackContent {
                 QuackText(
@@ -73,7 +83,7 @@ class TextTest {
     // - `Modifier.highlight`로 특정 텍스트에 [SpanStyle] 및 onClick 이벤트를 설정할 수 있음
     // onClick 이벤트 작동을 테스트함
     @Test
-    fun Set_SpanStyle_And_OnClickEvent_With_ModifierHighlight() {
+    fun set_SpanStyle_and_OnClickEvent_with_ModifierHighlight() {
         val oneOnClick: (String) -> Unit = mock()
         val twoOnClick: (String) -> Unit = mock()
         val threeOnClick: (String) -> Unit = mock()
@@ -81,7 +91,7 @@ class TextTest {
         composeTestRule.setQuackContent {
             QuackText(
                 modifier = Modifier
-                    .testTag("text")
+                    .markTest()
                     .highlight(
                         highlights = listOf(
                             "one" to oneOnClick, // 1~3
@@ -94,7 +104,7 @@ class TextTest {
             )
         }
 
-        composeTestRule.onQuackNodeWithTag("text").performTouchInput {
+        composeTestRule.onTest().performTouchInput {
             immediateXMoveTo(2f) // one
             immediateXMoveTo(6f) // two
             immediateXMoveTo(11f) // three
@@ -109,13 +119,13 @@ class TextTest {
 
     // - `Modifier.highlight`로 특정 텍스트에 [SpanStyle] 및 광역 onClick 이벤트를 설정할 수 있음
     @Test
-    fun Set_SpanStyle_And_GlobalOnClickEvent_With_ModifierHighlight() {
+    fun set_SpanStyle_and_GlobalOnClickEvent_with_ModifierHighlight() {
         val globalOnClick: (String) -> Unit = mock()
 
         composeTestRule.setQuackContent {
             QuackText(
                 modifier = Modifier
-                    .testTag("text")
+                    .markTest()
                     .highlight(
                         texts = listOf("one", "two", "three"),
                         globalOnClick = globalOnClick,
@@ -125,10 +135,53 @@ class TextTest {
             )
         }
 
-        composeTestRule.onQuackNodeWithTag("text").performClick()
+        composeTestRule.onTest().performClick()
 
         composeTestRule.runOnIdle {
             verify(globalOnClick, times(1)).invoke(any())
         }
     }
+
+    // `Modifier#span`과 `buildAnnotatedString`은 같은 텍스트 비쥬얼을 그림
+    @Test
+    fun ModifierSpan_and_buildAnnotatedString_rendered_same_ui() {
+        composeTestRule.setQuackContent {
+            Column {
+                QuackText(
+                    modifier = Modifier
+                        .markTest()
+                        .span(
+                            texts = listOf("@성빈", "코루틴"),
+                            style = SpanStyle(
+                                color = QuackColor.DuckieOrange.value,
+                                fontWeight = FontWeight.SemiBold,
+                            ),
+                        ),
+                    text = "@성빈 코루틴 잘 쓰고 싶다.",
+                    typography = QuackTypography.Body1,
+                )
+                BasicText(
+                    modifier = Modifier.markGolden(),
+                    text = buildAnnotatedString {
+                        val style = SpanStyle(
+                            color = QuackColor.DuckieOrange.value,
+                            fontWeight = FontWeight.SemiBold,
+                        )
+                        withStyle(style) { append("@성빈") }
+                        append(" ")
+                        withStyle(style) { append("코루틴") }
+                        append(" 잘 쓰고 싶다.")
+                    },
+                    style = QuackTypography.Body1.asComposeStyle(),
+                    overflow = TextOverflow.Ellipsis,
+                )
+            }
+        }
+
+        val test = composeTestRule.onTest().captureToBitmap()
+        val golden = composeTestRule.onGolden().captureToBitmap()
+
+        screenshotTest(name = "Modifier#span", test = test, golden = golden)
+    }
 }
+
