--- core/src/test/kotlin/team/duckie/quackquack/core/component/README.md
@@ -1,3 +1,0 @@
-## test/component
-
-이 모듈은 UI 컴포넌트의 비쥬얼을 테스트합니다. [paparazzi](https://github.com/cashapp/paparazzi)로 진행됩니다.
